requirements.txt
pandas
numpy
streamlit_option_menu